+++ src/main/java/roomescape/config/ReservationServiceConfig.java
@@ -3,8 +3,8 @@ package roomescape.config;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import roomescape.repository.reservation.ReservationRepository;
-import roomescape.repository.reservationtime.ReservationTimeRepository;
+import roomescape.persistence.repository.reservation.ReservationRepository;
+import roomescape.persistence.repository.reservationtime.ReservationTimeRepository;
 import roomescape.service.ReservationService;
 
 @Configuration
