+++ src/main/java/roomescape/console/ConsoleApplication.java
@@ -1,9 +1,9 @@
-package roomescape.application;
+package roomescape.console;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import roomescape.controller.ConsoleController;
+import roomescape.console.presentation.controller.ConsoleController;
 
 @Profile("console")
 @Component
