+++ src/main/java/roomescape/presentation/dto/CreateReservationDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.presentation.dto;
 
 public record CreateReservationDto(String name, String date, Long timeId) {
 }
