+++ src/main/java/roomescape/presentation/dto/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.presentation.dto;
 
 import roomescape.entity.Reservation;
 import roomescape.util.DateTimeFormatUtils;
